dxf: add dxf overlay with utm conversion. (line work only)

+++ Utilities/AP_GPS_SBF.cs
@@ -332,7 +332,7 @@ namespace MissionPlanner.Utilities
 
                     state.have_vertical_velocity = true;
 
-                    float ground_vector_sq = state.velocity[0]*state.velocity[0] + state.velocity[1]*state.velocity[1];
+                    float ground_vector_sq = (float)(state.velocity[0]*state.velocity[0] + state.velocity[1]*state.velocity[1]);
                     state.ground_speed = (float) safe_sqrt(ground_vector_sq);
 
                     state.ground_course = (float) (ToDeg(atan2f(state.velocity[1], state.velocity[0])));
